+++ packages/eslint-config-godaddy-typescript/CHANGELOG.md
@@ -1,5 +1,16 @@
 # eslint-config-godaddy-typescript
 
+## 6.0.0
+
+### Major Changes
+
+- 0fabf67: Upgrade to ESLint 10
+
+### Patch Changes
+
+- Updated dependencies [0fabf67]
+  - eslint-config-godaddy@9.0.0
+
 ## 5.0.0
 
 ### Major Changes